@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-redux": "^7.2.6",
-    "react-scripts": "^5.0.0",
+    "@reduxjs/toolkit": "^1.8.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
+    "react-redux": "^8.0.4",
+    "react-scripts": "^5.0.1",
     "react-tooltip": "^4.2.21",
-    "recharts": "^2.1.9",
+    "recharts": "^2.1.14",
     "redux-thunk": "^2.4.1",
-    "serve": "^13.0.2",
-    "source-map-explorer": "^2.5.2",
+    "serve": "^14.0.1",
+    "source-map-explorer": "^2.5.3",
     "use-sound": "^4.0.1",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^3.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -46,6 +46,6 @@
     ]
   },
   "devDependencies": {
-    "prettier": "2.5.1"
+    "prettier": "2.7.1"
   }
 }
