@@ -1,13 +1,14 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import Home from "./Pages/Home";
 import { Provider } from "react-redux";
 import store from "./State/Store";
 import "./index.css";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
   <Provider store={store}>
     <Home />
-  </Provider>,
-  document.getElementById("root")
+  </Provider>
 );
